Refactors document-worksapce to separate out entire workspace concerns and document workspace concerns
