+++ src/components/document/collapsed-document-workspace-tab.scss
