+++ src/components/document/document-workspace.sass
@@ -1,60 +1,59 @@
 @import ../../components/vars
 
-.document-workspace
-  .drag-handler
+.drag-handler
+  position: absolute
+  top: $header-height
+  left: 0
+  right: 0
+  bottom: 0
+
+.single-workspace
+  position: absolute
+  top: $header-height
+  left: calc(#{$collapsed-tab-width} + 4px)
+  right: 0
+  bottom: $half-nav-expanded-peek
+  height: calc(100vh - (#{$header-height} + #{$workspace-content-margin} * 2))
+  margin-top: $workspace-content-margin
+  transition: .5s
+  &.half
+    left: 50%
+
+.left-workspace
+  position: absolute
+  top: $header-height + $half-nav-expanded-peek
+  left: $left-nav-contracted-width + $nav-expanded-peek + $workspace-toolbar-width
+  right: 50%
+  bottom: $half-nav-expanded-peek
+
+  .workspace
+    margin-right: $half-margin
+
+.right-workspace
+  position: absolute
+  top: $header-height + $half-nav-expanded-peek
+  left: 50%
+  right: $right-nav-contracted-width + $nav-expanded-peek
+  bottom: $half-nav-expanded-peek
+
+  .workspace
+    margin-left: $half-margin
+
+  .comparison-placeholder
     position: absolute
-    top: $header-height
+    top: 0
     left: 0
     right: 0
     bottom: 0
-
-  .single-workspace
-    position: absolute
-    top: $header-height
-    left: calc(#{$collapsed-tab-width} + 4px)
-    right: 0
-    bottom: $half-nav-expanded-peek
-    height: calc(100vh - (#{$header-height} + #{$workspace-content-margin} * 2))
-    margin-top: $workspace-content-margin
-    transition: .5s
-    &.half
-      left: 50%
-
-  .left-workspace
-    position: absolute
-    top: $header-height + $half-nav-expanded-peek
-    left: $left-nav-contracted-width + $nav-expanded-peek + $workspace-toolbar-width
-    right: 50%
-    bottom: $half-nav-expanded-peek
-
-    .workspace
-      margin-right: $half-margin
-
-  .right-workspace
-    position: absolute
-    top: $header-height + $half-nav-expanded-peek
-    left: 50%
-    right: $right-nav-contracted-width + $nav-expanded-peek
-    bottom: $half-nav-expanded-peek
-
-    .workspace
-      margin-left: $half-margin
-
-    .comparison-placeholder
-      position: absolute
-      top: 0
-      left: 0
-      right: 0
-      bottom: 0
-      color: #000
-      background-color: #F8FCFE
-      margin: $margin
-      border: $border-width solid $border-color
-      border-radius: $border-radius
-      display: flex
-      justify-content: center
-      align-items: center
-      text-align: center
+    color: #000
+    background-color: #F8FCFE
+    margin: $margin
+    border: $border-width solid $border-color
+    border-radius: $border-radius
+    display: flex
+    justify-content: center
+    align-items: center
+    text-align: center
 
 .dialog
   &.create-personal
