+++ src/components/navigation/nav-tab-panel.sass
@@ -4,18 +4,17 @@
   display: flex
   flex-direction: row
   justify-content: flex-start
-  width: 50%
   height: calc(100vh - (#{$header-height} + #{$workspace-content-margin} * 2))
   position: fixed
+  width: 0px
   left: 0px
   top: $header-height
   transition-duration: .5s
   margin-top: $workspace-content-margin
   background-color: white
   border-radius: 0 6px 0 0
-  visibility: hidden
   &.shown
-    visibility: visible
+    width: 50%
   &.half
     width: calc(50% - 4px)
   &.full
