+++ src/components/workspace/workspace.sass
@@ -1,0 +1,12 @@
+@import ../../components/vars
+
+.workspace
+  display: flex
+  flex-direction: row
+
+  .drag-handler
+    position: absolute
+    top: $header-height
+    left: 0
+    right: 0
+    bottom: 0
